painelsaude/requirements.txt
aniso8601==9.0.1
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
DateTime==4.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Faker==18.11.2
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
mock==5.0.2
numpy==1.24.4
openpyxl==3.0.9
packaging==23.1
pandas==2.0.2
pluggy==1.2.0
psycopg==3.1.9
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
pytest==7.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
tomli==2.0.1
tqdm==4.65.0
typing-extensions==4.6.3
tzdata==2023.3
uuid==1.30
Werkzeug==2.0.2
zope.interface==5.4.0